refactor: 불필요한 로직 및 상수 삭제
